+++ miles/backends/training_utils/loss.py
@@ -8,8 +8,8 @@ from miles.backends.training_utils.loss_hub import get_loss_function
 from miles.backends.training_utils.loss_hub.advantages import compute_advantages, normalize_advantages
 from miles.backends.training_utils.loss_hub.base_types import LossFnInput
 from miles.backends.training_utils.loss_hub.logit_processors import get_log_probs_and_entropy, get_values
+from miles.backends.training_utils.loss_hub.math_utils import compute_approx_kl
 from miles.backends.training_utils.parallel import get_parallel_state
-from miles.utils.ppo_utils import compute_approx_kl
 from miles.utils.types import RolloutBatch
 
 __all__ = [
