+++ miles/backends/training_utils/loss_hub/logit_processors.py
@@ -4,8 +4,8 @@ from collections.abc import Iterator
 import torch
 
 from miles.backends.training_utils.cp_utils import allgather_cp_redistribute, get_logits_and_tokens_offset_with_cp
+from miles.backends.training_utils.loss_hub.math_utils import calculate_log_probs_and_entropy
 from miles.backends.training_utils.parallel import get_parallel_state
-from miles.utils.ppo_utils import calculate_log_probs_and_entropy
 
 
 def get_responses(
